feat: create hotel provider
